[Tools] The generator's store rows carry the store's own types
storeEffect's Type and Aura are dbcenums.SpellEffectType and EffectAuraType, the class flags
of a row and an effect are core.ClassFlags, and a row's school is a core.SpellSchool, so the
generator's mirror drops storeClassFlags and the casts around the effect and aura checks. The
captured inputs encode the same numbers and the rendered store is unchanged.

+++ tools/database/spelldata_closure.go
@@ -226,7 +226,7 @@ func spellEdges(t *spellTables, id int32) []int32 {
 		if e.TriggerID > 0 {
 			next = append(next, e.TriggerID)
 		}
-		if e.Aura == int32(dbcenums.A_OVERRIDE_ACTIONBAR_SPELLS) && e.BasePoints > 0 {
+		if e.Aura == dbcenums.A_OVERRIDE_ACTIONBAR_SPELLS && e.BasePoints > 0 {
 			next = append(next, int32(e.BasePoints))
 		}
 	}
